voice-ai-os/voice-core/pyproject.toml
[project]
name = "voice-core"
version = "0.1.0"
description = "Voice Core (Layer 1) - Immutable voice AI foundation using Pipecat framework"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "SpotFunnel AI"}
]
license = {text = "MIT"}
keywords = ["voice-ai", "pipecat", "stt", "tts", "llm"]

dependencies = [
    "pipecat-ai[daily,deepgram,openai,elevenlabs]>=0.0.46",
    "daily-python>=0.10.1",
    "deepgram-sdk>=3.2.7",
    "openai>=1.12.0",
    "elevenlabs>=0.2.26",
    "python-dotenv>=1.0.1",
    "aiohttp>=3.9.1",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.4",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "black>=24.0.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["voice_core"]

[tool.setuptools.package-data]
voice_core = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
